Implement data transformation for UI display
This commit introduces the addition of a handler and related data class to facilitate the transformation of raw data objects received from the network into objects suitable for UI display. The handler, along with the data class, enables the conversion and preparation of data in a format optimised for rendering in the user interface.